agent-hiring-mvp/agenthub-sdk/templates/rag_agent/requirements.txt
llama-index
llama-index-readers-web
llama-index-readers-file
llama-index-llms-openai
llama-index-embeddings-openai
python-dotenv
requests
beautifulsoup4
pypdf